Refactor: Add service.rs for systemd installation
